requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.0
elasticsearch>=8.11.0,<9.0.0
gunicorn>=21.0.0
